# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Arodona Jewelry Backend API",
    "version": "1.0.0",
    "description": "E-commerce API for Arodona Jewelry store"
  },
  "servers": [
    {
      "url": "http://localhost:5858",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/cart": {
      "get": {
        "summary": "Get current user's cart",
        "tags": ["Cart"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Cart retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartOut"
                }
              }
            }
          }
        }
      }
    },
    "/cart/add": {
      "post": {
        "summary": "Add item to cart",
        "tags": ["Cart"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item added to cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/cart/clear": {
      "post": {
        "summary": "Clear cart",
        "tags": ["Cart"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Cart cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CartItemCreate": {
        "type": "object",
        "properties": {
          "product": {
            "type": "string"
          },
          "variant_id": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": ["product", "quantity"]
      },
...
      "CartItemOut": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string"
          },
          "variant_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "quantity": {
            "type": "integer"
          },
          "item_total": {
            "type": "number"
          }
        },
        "required": ["product_id", "name", "price", "quantity", "item_total"]
      },
      "CartOut": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemOut"
            }
          },
          "subtotal": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "tax": {
            "type": "number"
          },
          "delivery_fee": {
            "type": "number"
          },
          "total_amount": {
            "type": "number"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["user_id", "items", "subtotal", "discount", "tax", "delivery_fee", "total_amount", "updated_at", "created_at"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}